Maywood: enable theme default
For every font-family defined, Maywood provides a fallback to the
CSS variable (--font-base or --font-headings). It doesn't make use
of the --font-base-default and --font-headings-default variables,
so we can get rid of them. We, instead, need an alternative way
to declare this theme supports the "Theme Default" option,
hence the enable-theme-default.