Fix #724 : Lodestar gallery block margin issue
This solves the bottom-margin issue with the gallery block on the front page of Lodestar. The theme wants to set a margin-bottom for the last item in the entry-content to be 0, but doesn't account for the block editor. This PR fixes this by limiting this rule to the last direct descendent of the entry-content.