Defaulted the color of input/textarea elements
The color of input and textarea elements were inheriting, which worked
in most situations.  However when nested in other elements (such as a
cover block) the color was reset to white which caused the form elements
text to be unreadable.

Since this theme doesn't allow for text color changes the color value of
text elements is set to the same value defaulting all text elements.