Remove 'dark-editor-style' theme support
This fixes the issue where some styles (such as the image caption) have
"dark mode" support added to them.  This is fine if a user has "dark
mode" enabled for their system.  But in the default mode it isn't a
"dark" style and therefore those stylistic behaviors introduce problems.

I did not notice any negative behavior due to this being removed.

This behavior seems to have been introduced due to recent chagnes in
Gutenberg that introduce additional editor styles for themes that
express support for 'dark-editor-style'.