Change CSS variable assignment order
First Seedlet was changed so that all custom CSS variables for color
definitions are associated with the 'seedlet-custom-color-overrides'
stylesheed.  (Previously only editor override styles were associated
with that stylesheet.

Next Spearhead was changed to first dequeue that stylesheet, THEN assign
the spearhead-specific style variables THEN re-enqueue that
seedlet-custom-color-overrides stylesheet.

This puts the customizer color variable assignments AFTER the Spearhead
variable assignments so that those are the values that are used.