Improved sandbox script to delete appropriately
This change collects a list of what files have been deleted in the
repository and removes them from the sandbox via rm.

If those files have since returned (deleted but returned) they will
still be deleted but, since those files will be in your local copy will
be synced to the sandbox.