Separate background & color
This changese creates a more specific selector for applying the color of
an element with a background color applied.

Previously the class that applied the background color also applied a
reasonable color color.

With this change only matches that also do not also supply a
'has-text-color' class will set a color property for a .has-*-background