Added tools to import/export content
This change adds tools to import/export content from the local wp-env
instance.

Content can be pulled from your sandbox, themedemo.wordpress.com or from
the theme-unit-test repository

The sandbox.sh script was changed to power the "export" (from sandbox or
themedemo) functionality.

The .wp-env.json was changed to supply the content from theme-unit-test
repository, map local folders for import/export of the rest of the
resources and to provide the import plugin by default.

npm scripts have been added to package.json in order to drive the
commands.

Documentation in readme has been updated to reflect the additional
capabilities.