Proposal: automatically add strict types
This PR is an experiment that aims to check if it would be practical to
automatically add strict types to all PHP files that are missing them.
Initially, it will add them to _all_ PHP files, but con subsequen runs,
it'll only affect new files.

The strict types declaration has become a requirement on WordPress.com
in order to commit new PHP files to the codebase, and solving this
upstream (eg. on Create Block Theme) might not be the best, since
prescribing strict types to its broad userbase could prove problematic.