Remove copyright statement (#7988)
The footer copyright statement has been removed from the following themes:

 * Vitrum
 *  Fontaine (x2)
 *  Pendant
 *  Bitacora
 *  Jaida

The standard for a while has been to use "Designed with WordPress" (or something similar) in the footer, rather than a copyright statement. Additionally the copyright line is often incorrect - where it hardcodes the year, this is obviously not accurate and where it credits the name of the blog, or states "All Rights Reserved", this may not be accurate either. We should leave it to users to set their own copyright statements appropriately, and most of these themes leave an empty paragraph available to make it easy to do that.